SFEperl-podlators: bump to 4.09, package for perl-522 is added and for perl-520 is obsolete
--HG--
branch : spec_SFEperl-podlators

+++ specs/SFEperl-podlators.spec
@@ -4,8 +4,8 @@
 %define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define include_executable 0
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define include_executable 1
 
 %define cpan_name podlators
 %define sfe_cpan_name podlators
@@ -14,11 +14,11 @@
 Summary:               Convert POD data to various other formats
 Name:                  SFEperl-%{sfe_cpan_name}
 IPS_package_name:      library/perl-5/%{ips_cpan_name}
-Version:               2.5.3
-IPS_component_version: 2.5.3
+Version:               4.09
+IPS_component_version: 4.9
 License:               perl_5
 URL:                   https://metacpan.org/pod/podlators
-Source0:               http://cpan.metacpan.org/authors/id/R/RR/RRA/podlators-2.5.3.tar.gz
+Source0:               http://cpan.metacpan.org/authors/id/R/RR/RRA/podlators-%{version}.tar.gz
 BuildRoot:             %{_tmppath}/%{name}-%{version}-build
 
 %description
@@ -30,10 +30,11 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-584
 Summary:          Convert POD data to various other formats
 BuildRequires:    runtime/perl-584 = *
 BuildRequires:    library/perl-5/extutils-makemaker-584
+BuildRequires:    library/perl-5/encode-584
+BuildRequires:    library/perl-5/pod-simple-584
 Requires:         runtime/perl-584 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/encode-584
-Requires:         library/perl-5/pathtools-584
 Requires:         library/perl-5/pod-simple-584
 
 %description 584
@@ -46,10 +47,11 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-510
 Summary:          Convert POD data to various other formats
 BuildRequires:    runtime/perl-510 = *
 BuildRequires:    library/perl-5/extutils-makemaker-510
+BuildRequires:    library/perl-5/encode-510
+BuildRequires:    library/perl-5/pod-simple-510
 Requires:         runtime/perl-510 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/encode-510
-Requires:         library/perl-5/pathtools-510
 Requires:         library/perl-5/pod-simple-510
 
 %description 510
@@ -62,10 +64,11 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-512
 Summary:          Convert POD data to various other formats
 BuildRequires:    runtime/perl-512 = *
 BuildRequires:    library/perl-5/extutils-makemaker-512
+BuildRequires:    library/perl-5/encode-512
+BuildRequires:    library/perl-5/pod-simple-512
 Requires:         runtime/perl-512 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/encode-512
-Requires:         library/perl-5/pathtools-512
 Requires:         library/perl-5/pod-simple-512
 
 %description 512
@@ -78,36 +81,39 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-516
 Summary:          Convert POD data to various other formats
 BuildRequires:    runtime/perl-516 = *
 BuildRequires:    library/perl-5/extutils-makemaker-516
+Requires:         library/perl-5/%{ips_cpan_name}
+BuildRequires:    library/perl-5/encode-516
+BuildRequires:    library/perl-5/pod-simple-516
 Requires:         runtime/perl-516 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/encode-516
-Requires:         library/perl-5/pathtools-516
 Requires:         library/perl-5/pod-simple-516
 
 %description 516
 Convert POD data to various other formats
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{ips_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          Convert POD data to various other formats
-BuildRequires:    runtime/perl-520 = *
-BuildRequires:    library/perl-5/extutils-makemaker-520
-Requires:         runtime/perl-520 = *
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/extutils-makemaker-522
+BuildRequires:    library/perl-5/encode-522
+BuildRequires:    library/perl-5/pod-simple-522
+Requires:         runtime/perl-522 = *
 Requires:         library/perl-5/%{ips_cpan_name}
-Requires:         library/perl-5/encode-520
-Requires:         library/perl-5/pathtools-520
-Requires:         library/perl-5/pod-simple-520
+Requires:         library/perl-5/encode-522
+Requires:         library/perl-5/pod-simple-522
 
-%description 520
+%description 522
 Convert POD data to various other formats
 %endif
 
 
 %prep
 %setup -q -n %{cpan_name}-%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
@@ -120,8 +126,17 @@ build_with_makefile.pl_for() {
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
                    LIB=${vendor_dir}
-    make
-    [ x${test} = 'xwithout_test' ] || make test
+
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null
+    if [ $? -eq 0 ]
+    then
+        make CC='cc -m32' LD='cc -m32'
+        [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC='cc -m32' LD='cc -m32'
+    else
+        make CC='cc -m64' LD='cc -m64'
+        [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC='cc -m64' LD='cc -m64'
+    fi
+
     make pure_install
 }
 
@@ -136,16 +151,26 @@ build_with_build.pl_for() {
                    --installdirs vendor \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    [ x${test} = 'xwithout_test' ] || ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
+    ${bindir}/perl ./Build clean
 }
 
 modify_bin_dir() {
     perl_ver=$1
     if [ -d $RPM_BUILD_ROOT/usr/bin ]
     then
-       # perl-5{12,16,20} provide bin/{pod2man,pod2txt}
-       rm -rf $RPM_BUILD_ROOT/usr/bin
+      [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+      mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
+    fi
+      
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin ]
+    then
+        for i in $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin/*
+        do
+            sed -i.bak -e "s!/usr/bin/env perl!/usr/perl5/${perl-ver}/bin/perl!" ${i}
+            [ -f ${i}.bak] || rm -f ${i}.bak
+        done
     fi
 }
 
@@ -161,7 +186,11 @@ modify_man_dir() {
             mv $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man $RPM_BUILD_ROOT%{_datadir}/
             rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
         fi
-        rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        if [ %{include_executable} -eq 0 ]
+        then
+            rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        fi
+
     fi
 }
 
@@ -196,8 +225,8 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
 %endif
 
 %install
@@ -258,17 +287,18 @@ rm -rf %{buildroot}
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+/usr/perl5/vendor_perl/5.22
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
-
 
 %changelog
+* Wed Apr 26 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- bump to 4.09, package for perl-522 is added and for perl-520 is obsolete
 * Sun Nov 15 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - initial commit
